refactor(tests): move each test group into a separate file
